Declaratively disable spring-boot autoconfigurations
Remove `exclude = {class list}` from `@EnableAutoConfiguration` in
service configuration classes, and disable the spring-boot
auto-configurations that cuase trouble declaratively in bootstrap.yml
instead.
