--- services/wcs/src/main/resources/bootstrap-test.yml
@@ -1,12 +1,0 @@
-spring:
-  application:
-    name: wcs-service
-  cloud:
-    config:
-      enabled: false
-      discovery:
-        enabled: false
-server:
-  port: 9091
-  servlet:
-    context-path: /
