+++ services/web-ui/src/main/java/org/geoserver/cloud/webui/WebUIApplication.java
@@ -14,7 +14,6 @@ public class WebUIApplication {
         try {
             SpringApplication.run(WebUIApplication.class, args);
         } catch (RuntimeException e) {
-            e.printStackTrace();
             System.exit(-1);
         }
     }
