+++ services/wfs/src/test/java/org/geoserver/cloud/wfs/WfsApplicationTest.java
@@ -1,0 +1,22 @@
+/*
+ * (c) 2020 Open Source Geospatial Foundation - all rights reserved This code is licensed under the
+ * GPL 2.0 license, available at the root application directory.
+ */
+package org.geoserver.cloud.wfs;
+
+import org.junit.Test;
+import org.junit.runner.RunWith;
+import org.springframework.boot.autoconfigure.EnableAutoConfiguration;
+import org.springframework.boot.test.context.SpringBootTest;
+import org.springframework.test.context.ActiveProfiles;
+import org.springframework.test.context.junit4.SpringRunner;
+
+@RunWith(SpringRunner.class)
+@SpringBootTest(classes = WfsApplication.class)
+@EnableAutoConfiguration
+@ActiveProfiles("test")
+public class WfsApplicationTest {
+
+    @Test
+    public void contextLoads() {}
+}
