+++ services/wms/src/test/java/org/geoserver/cloud/wms/WmsApplicationTest.java
@@ -2,7 +2,7 @@
  * This code is licensed under the GPL 2.0 license, available at the root
  * application directory.
  */
-package org.geoserver.cloud.restconfig;
+package org.geoserver.cloud.wms;
 
 import org.junit.Test;
 import org.junit.runner.RunWith;
@@ -13,7 +13,7 @@ import org.springframework.test.context.junit4.SpringRunner;
 @RunWith(SpringRunner.class)
 @SpringBootTest
 @ActiveProfiles("test")
-public class RestConfigApplicationTests {
+public class WmsApplicationTest {
 
     @Test
     public void contextLoads() {}
