+++ services/wps/src/main/java/org/geoserver/cloud/wps/WpsApplicationConfiguration.java
@@ -2,7 +2,7 @@
  * (c) 2020 Open Source Geospatial Foundation - all rights reserved This code is licensed under the
  * GPL 2.0 license, available at the root application directory.
  */
-package org.geoserver.cloud.wcs;
+package org.geoserver.cloud.wps;
 
 import org.geoserver.cloud.config.factory.FilteringXmlBeanDefinitionReader;
 import org.springframework.context.annotation.Configuration;
@@ -12,10 +12,10 @@ import org.springframework.context.annotation.ImportResource;
 @ImportResource( //
     reader = FilteringXmlBeanDefinitionReader.class, //
     locations = { //
-        "jar:gs-wcs-.*!/applicationContext.xml", //
-        "jar:gs-wcs1_0-.*!/applicationContext.xml", //
-        "jar:gs-wcs1_1-.*!/applicationContext.xml", //
-        "jar:gs-wcs2_0-.*!/applicationContext.xml" //
+        "jar:gs-wps-.*!/applicationContext.xml" // , //
+        // // REVISIT: wps won't start without the web components! see note in pom.xml
+        // "jar:gs-web-core-.*!/applicationContext.xml", //
+        // "jar:gs-web-wps-.*!/applicationContext.xml", //
     }
 )
-public class WcsApplicationConfiguration {}
+public class WpsApplicationConfiguration {}
